refactor: BE architecture & implementing DB transactions
