+++ backend-sb/.env-example
@@ -9,7 +9,6 @@ DB_PORT=
 DB_NAME=
 DB_USERNAME=
 DB_PASSWORD=
-DB_SYNCHRONIZE=
 
 # Redis setup
 RDS_HOST=
