+++ backend-sb/.gitignore
@@ -17,6 +17,7 @@
 
 .yarn
 .cache
+ormconfig.json
 
 npm-debug.log*
 yarn-debug.log*
