+++ backend-sb/package.json
@@ -1,8 +1,8 @@
 {
   "name": "backend-sb",
   "version": "0.0.1",
-  "description": "",
-  "author": "",
+  "description": "SonicBoom Backend Server",
+  "author": "Cuong Nguyen <cuong.nugi@gmail.com>",
   "license": "MIT",
   "scripts": {
     "prebuild": "rimraf dist",
@@ -12,6 +12,12 @@
     "start:dev": "cross-env NODE_ENV=development nest start --watch",
     "start:debug": "cross-env NODE_ENV=development nest start --debug --watch",
     "start:prod": "cross-env NODE_ENV=production node dist/main",
+    "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js",
+    "db:setup": "(rimraf ormconfig.json || :) && ts-node -r tsconfig-paths/register scripts/generate-typeorm-config.ts",
+    "db:generate": "yarn typeorm migration:generate -n",
+    "db:migrate": "yarn typeorm migration:run",
+    "preprod:typeorm": "(rimraf ormconfig.json || :) && node dist/scripts/generate-typeorm-config.js",
+    "prestart:prod": "yarn preprod:typeorm && node ./node_modules/typeorm/cli.js migration:run",
     "lint": "tslint -p tsconfig.json -c tslint.json",
     "test": "jest",
     "test:watch": "jest --watch",
@@ -37,7 +43,7 @@
     "express-rate-limit": "5.1.1",
     "helmet": "3.22.0",
     "lodash": "4.17.19",
-    "nestjs-redis": "1.2.7",
+    "nestjs-redis": "1.2.8",
     "passport": "0.4.1",
     "passport-jwt": "4.0.0",
     "passport-spotify": "1.1.0",
