+++ backend-sb/src/auth/decorators/get-user.decorator.ts
@@ -1,6 +1,8 @@
 import { createParamDecorator } from '@nestjs/common';
 import { User } from '../entities/user.entity';
 
-export const GetUser = createParamDecorator((data, req): User | undefined => {
-  return req.args[0].user as User;
-});
+export const GetUser = createParamDecorator(
+  (data, req): User => {
+    return req.args[0].user as User;
+  }
+);
