+++ backend-sb/src/auth/interfaces/auth.enum.ts
