+++ backend-sb/src/auth/interfaces/spotifyPayload.interface.ts
