+++ backend-sb/src/auth/strategies/spotify.strategy.ts
@@ -3,9 +3,9 @@ import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-spotify';
 import { AuthService } from '../auth.service';
 import mainConfig from '../../config/main.config';
-import { UserDataInterface } from '../../interfaces/user-data.interface';
+import { UserDataInterface } from '../interfaces/user-data.interface';
 import { User } from '../entities/user.entity';
-import { UrlEnums } from '../enums/urls.enum';
+import { UrlEnums } from '../interfaces/urls.enum';
 
 @Injectable()
 export class SpotifyStrategy extends PassportStrategy(Strategy, 'spotify') {
