+++ backend-sb/src/config/mainConfig.interface.ts
@@ -1,3 +1,5 @@
+import { TypeOrmModuleOptions } from '@nestjs/typeorm';
+
 export interface MainConfigInterface {
   serverSettings: {
     serverMode: string;
